+++ server/service_lock.go
@@ -1,0 +1,9 @@
+package server
+
+type lockService interface {
+	init() error
+	tryLock(resource string) error
+	release(resource string) error
+	exists(resource string) bool
+	stop()
+}